ml-service/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
joblib
requests
pydantic
sentence-transformers
faiss-cpu
openai